requirements.txt
requests==2.32.3
schedule==1.2.2
python-dateutil==2.9.0
